Break out encrypt ami into own module
Allows for library usage.  Also, provides a better organization
for future functionality.

Figure out version in setup.py
