+++ brkt_cli/encrypt_ami_args.py
@@ -1,0 +1,46 @@
+def setup_encrypt_ami_args(parser):
+    parser.add_argument(
+        'ami',
+        metavar='AMI_ID',
+        help='The AMI that will be encrypted'
+    )
+    parser.add_argument(
+        '--encrypted-ami-name',
+        metavar='NAME',
+        dest='encrypted_ami_name',
+        help='Specify the name of the generated encrypted AMI',
+        required=False
+    )
+    parser.add_argument(
+        '--encryptor-ami',
+        metavar='ID',
+        dest='encryptor_ami',
+        help='Bracket Encryptor AMI',
+        required=False
+    )
+    parser.add_argument(
+        '--key',
+        metavar='NAME',
+        help='EC2 SSH Key Pair name',
+        dest='key_name',
+        required=True
+    )
+    parser.add_argument(
+        '--validate-ami',
+        dest='no_validate_ami',
+        action='store_true',
+        help="Validate AMI properties (default)"
+    )
+    parser.add_argument(
+        '--no-validate-ami',
+        dest='no_validate_ami',
+        action='store_false',
+        help="Don't validate AMI properties"
+    )
+    parser.add_argument(
+        '--region',
+        metavar='NAME',
+        help='AWS region (e.g. us-west-2)',
+        dest='region',
+        required=True
+    )
