+++ brkt_cli/service.py
@@ -15,6 +15,7 @@
 import abc
 import re
 import boto
+import boto.sts
 import logging
 import json
 import urllib2
@@ -193,6 +194,17 @@ class AWSService(BaseAWSService):
         self.region = region
         self.key_name = key_name
         self.conn = boto.vpc.connect_to_region(region)
+
+    def connect_as(self, role, region, session_name):
+        sts_conn = boto.sts.connect_to_region(region)
+        creds = sts_conn.assume_role(role, session_name)
+        conn = boto.vpc.connect_to_region(
+            region,
+            aws_access_key_id=creds.credentials.access_key,
+            aws_secret_access_key=creds.credentials.secret_key,
+            security_token=creds.credentials.session_token)
+        self.region = region
+        self.conn = conn
 
     def run_instance(self,
                      image_id,
